\name{colnames-methods}
\docType{methods}
\alias{colnames-methods}
\alias{colnames,SGOL-method}
\alias{rownames-methods}
\alias{rownames,SGOL-method}
\alias{[-methods}
\alias{[,SGOL-method}
\title{Methods for Function colnames/rownames in Package `cghMCR'}
\description{
 Methods for function \code{colnames} or \code{rownames} in Package `cghMCR' to extract row or column names from a SGOL object
}
\section{Methods}{
\describe{
\item{x = "SGOL"}{Extract colnum/row names from SGOL object}
}}
\keyword{methods}
