\name{plot.DNAcopy}
\alias{plot.DNAcopy}
\title{A function to plot the original data along with the segments
  identified using segment of DNAcopy.}
\description{
  This function takes a \code{\link{DNAcopy}} object generated by
  running \code{\link{getSegments}} and then plots the original data
  along with the segments identified. 
}
\usage{
plot.DNAcopy(x, ...)
}
\arguments{
  \item{x}{\code{x} is a \code{\link{DNAcopy}} object
    generated by \code{\link{getSegments}}}
  \item{\dots}{\code{\dots} may contain an optional boolean for save to
    indicate whether to show the plot on the screen (FALSE. Default) or
    return the name of the plot (TRUE) to allow the users to save the plot}
}
\details{
  This function only works with the \code{\link{DNAcopy}} returned by
  \code{\link{getSegments}}. It will work with the \code{\link{DNAcopy}}
  returned by \code{\link{segment}} of the DNAcopy package when the a
  column containing the probe ids is added as the first column of the
  "data" element of the object.
}
\value{
  This function returns invisible() or the name of the plot depending on
  whether save is set to FALSE of TRUE.
}
\references{The DNAcopy package}
\author{Jianhua Zhang}
\note{The function is a contribution of The Center for Applied Cancer
  Science of Dana-Farber Cancer Institute}
\seealso{\code{\link{getSegments}}}
\examples{
  if(interactive()){
    require("marray")
    sampleData <- read.Agilent(list.files(file.path(.path.package("cghMCR"),
      "sampledata"), full.name = TRUE, pattern = "sample"), path = "")
    segments <- getSegments(sampleData)
    plot(segments)
  }
}
\keyword{hplot}
